requirements.txt
chromadb
pandas
llama-index
fastapi
sentence-transformers
torch
transformers
uvicorn
numpy
pyenchant
